tutorials/google-adk/05_parallel_agents/pyproject.toml
[project]
name = "travel-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.21.0",
]
